Fp8 Quantization Support (#62)
* small fixes

* initial commit

* bug fixes

* cleanup

* clarity comments

* clean up compression classes

* fixing zero point issues

* comment for hack

* update quant check

* cleanup fp8 dtypes

* cleanup

* clean up observer

* dtype fix

* docstrings

* fixes after rebase

* test fixes

* style

* get rid of broken segment

* fix broken code

+++ tests/test_compressors/test_pack_quant.py
@@ -110,10 +110,10 @@ def test_reload_match(tmp_path):
     dense_state_dict = {
         "dummy.weight": torch.rand((511, 350)),
         "dummy.weight_scale": torch.tensor(0.01, dtype=torch.float32),
-        "dummy.weight_zero_point": torch.tensor(0, dtype=torch.int32),
+        "dummy.weight_zero_point": torch.tensor(0, dtype=torch.int8),
         "dummy2.weight": torch.rand((128, 280)),
         "dummy2.weight_scale": torch.tensor(0.02, dtype=torch.float32),
-        "dummy2.weight_zero_point": torch.tensor(15, dtype=torch.int32),
+        "dummy2.weight_zero_point": torch.tensor(15, dtype=torch.int8),
     }
     quant_config = get_dummy_quant_config()
